99-com.rules: Combine ttyAMA[01] rules by using link_priority.
In 05cfe136f76677a891a0c77dc91d796913607c07, the ttyAMA[01] rule was split into two rules to support the bcm2711 (Raspberry Pi 4B, etc.) where ttyAMA0 and ttyAMA1 can both exist. This commit recombines the rules and instead uses link_priority to make the symlink prefer ttyAMA0. (This also has the advantage of not being dependent on /dev/ttyAMA0 being added before ttyAMA1.)

The priority of -100 is chosen for consistency with other rules (see `grep -r link_priority /lib/udev /etc/udev`).

Tested on a Raspberry Pi 400 with uart2 (ttyAMA1) enabled:
- `ls -l /dev/serial0` points to ttyAMA0.
- Switching line 17 to say "ttyAMA0" and rebooting causes the link to point to ttyAMA1.

Note that this can do the wrong thing if ttyAMA0 is not uart0, e.g. if uart0 is disabled but uart2 is enabled, or some other kernel/device-tree combination that causes a different uart to enumerate as uart0.